backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Werkzeug==3.0.1
python-dotenv
google-generativeai
snowflake-connector-python
scikit-learn
pandas